+++ src/harlequin/components/__init__.py
@@ -1,7 +1,7 @@
 from harlequin.components.code_editor import CodeEditor, EditorCollection
 from harlequin.components.data_catalog import CatalogItem, DataCatalog
 from harlequin.components.error_modal import ErrorModal
-from harlequin.components.export_screen import ExportScreen
+from harlequin.components.export_screen import ExportScreen, export_callback
 from harlequin.components.help_screen import HelpScreen
 from harlequin.components.results_viewer import ResultsViewer
 from harlequin.components.run_query_bar import RunQueryBar
@@ -11,6 +11,7 @@ __all__ = [
     "EditorCollection",
     "ErrorModal",
     "ExportScreen",
+    "export_callback",
     "HelpScreen",
     "ResultsViewer",
     "RunQueryBar",